requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
streamlit==1.36.0
faiss-cpu==1.8.0
scikit-learn==1.4.2
rank-bm25==0.2.2
markdown2==2.4.13
pypdf==4.2.0
openai==1.30.1
sentence-transformers==2.7.0
numpy==1.26.4
nltk==3.8.1
rouge-score==0.1.2
pydantic==2.7.1
pytest==8.2.0
requests==2.32.3
PyYAML==6.0.1
typer==0.12.3
Jinja2==3.1.4
matplotlib==3.9.0
flake8==7.1.0
